using TextMgr in client
+ fixed command list problem (delete while in use) -> light
+ chagne building light process

+++ Client/X-MACHINA/ClientNetwork/Contents/Script_RemotePlayer.h
@@ -55,7 +55,7 @@ private:
 
 	float mLerpVal = 0.f;
 
-	double mBezierTime = 0.f;
+	float mBezierTime = 0.f;
 
 	float t = 0;
